Added Comments and moved functions into appropriate areas
