+++ src/App.tsx
@@ -2,5 +2,6 @@ import React from "react";
 import StopWatch from "./StopWatch";
 
 export default function App() {
+  // Calls the object
   return <StopWatch />;
 }
