+++ src/StopWatch.tsx
@@ -1,12 +1,15 @@
 import React, { useEffect, useState } from "react";
 import StopWatchButton from "./StopWatchButton";
-import Styles from "./Styles";
+import Styles from "./Styles/Styles";
+import padNumber from "./Helper/PadNumbers";
 
 const StopWatch = () => {
+  // Used states to hold the variables that would change during the page
   const [time, setTime] = useState(0);
   const [isRunning, setRunning] = useState(false);
   const [laps, setLaps] = useState([]);
 
+  // Used the React useEffect to increment the timer
   useEffect(() => {
     if (isRunning) {
       let id = setInterval(() => setTime(time + 1));
@@ -14,6 +17,7 @@ const StopWatch = () => {
     }
   }, [time, isRunning]);
 
+  // Functions that are called when buttons are pressed
   const start = () => {
     setRunning(true);
   };
@@ -32,19 +36,22 @@ const StopWatch = () => {
     setLaps([...laps, { min: minutes, sec: seconds, ms: milliseconds }]);
   };
 
-  function padNumber(num: number, pad: number) {
-    return Math.floor(num).toString().padStart(pad, "0");
-  }
-
+  // Calculated the types of seconds based on the current timer
+  // Used a helper function to format the numbers correctly
   const milliseconds = padNumber(time % 100, 2);
   const seconds = padNumber((time % 6000) / 100, 2);
   const minutes = padNumber((time % 360000) / 6000, 2);
 
+  // I created a style class called Styles.tsx to hold all my css
+  // I wasn't sure if we were allowed to import libaries (such as styled-components)
+
   return (
     <div>
+      {/* The Timer */}
       <div style={Styles.timer}>
         {minutes}:{seconds}:{milliseconds}
       </div>
+      {/* The Buttons */}
       <div style={Styles.buttonGrid}>
         <StopWatchButton
           style={Styles.button}
@@ -71,7 +78,11 @@ const StopWatch = () => {
           disabled={!isRunning}
         />
       </div>
+
+      {/* The Laps */}
+      {/* I decided to only show the laps title if there was already a lap stored */}
       {laps.length > 0 && <div style={Styles.title}> Laps </div>}
+      {/* Creates a row for each lap that was stored */}
       {laps.length > 0 &&
         laps.map((l, index) => {
           return (
